requirements.txt
blinker==1.7.0
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==3.0.0
Flask-Cors==4.0.0
importlib-metadata==7.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
six==1.16.0
typing_extensions==4.9.0
tzdata==2023.3
Werkzeug==3.0.1
zipp==3.17.0
